% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsbm.R
\name{compute_final_stats}
\alias{compute_final_stats}
\title{Aggregate indicators into final statistics}
\usage{
compute_final_stats(ind)
}
\description{
Aggregate indicators into final statistics
}
\keyword{internal}
