% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsbm.R
\name{compute_indicators}
\alias{compute_indicators}
\title{Compute boolean/error indicators for evaluation}
\usage{
compute_indicators(rat, test, ratings)
}
\description{
Compute boolean/error indicators for evaluation
}
\keyword{internal}
