% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{compute_likelihood}
\alias{compute_likelihood}
\title{Compute log-likelihood of the model}
\usage{
compute_likelihood(data, theta, eta, pr)
}
\arguments{
\item{data}{Integer matrix (N x 3).}

\item{theta}{Numeric matrix (U x K).}

\item{eta}{Numeric matrix (I x L).}

\item{pr}{Numeric 3D array (K x L x R).}
}
\value{
Scalar log-likelihood value.
}
\description{
Compute log-likelihood of the model
}
\keyword{internal}
