% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsbm.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute all evaluation statistics from a prediction matrix}
\usage{
compute_stats(rat, test, ratings)
}
\description{
Compute all evaluation statistics from a prediction matrix
}
\keyword{internal}
