% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{create_encoding}
\alias{create_encoding}
\title{Create a fresh encoding environment}
\usage{
create_encoding()
}
\value{
An environment with NULL slots for obs_dict, items_dict, ratings_dict.
}
\description{
Create a fresh encoding environment
}
\keyword{internal}
