% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{normalize_with_d}
\alias{normalize_with_d}
\title{Normalise a matrix by pre-computed factors}
\usage{
normalize_with_d(mat, norm_factors)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{norm_factors}{Numeric matrix of the same shape.}
}
\value{
Normalised matrix.
}
\description{
Divides each row of \code{mat} by the corresponding row in \code{norm_factors}.
}
\keyword{internal}
