% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsbm.R
\name{predict.mmsbm}
\alias{predict.mmsbm}
\title{Predict ratings for new user-item pairs}
\usage{
\method{predict}{mmsbm}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted \code{mmsbm} object.}

\item{newdata}{A data.frame with three columns: users, items, ratings.}

\item{...}{Ignored (for S3 generic compatibility).}
}
\value{
A numeric matrix (N x R) of rating probability distributions.
}
\description{
Predict ratings for new user-item pairs
}
