% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{rename_values}
\alias{rename_values}
\title{Apply an encoding to a character vector}
\usage{
rename_values(x, dict)
}
\arguments{
\item{x}{Character vector.}

\item{dict}{Named integer vector (as returned by \code{create_values_dict}).}
}
\value{
Integer vector.
}
\description{
Apply an encoding to a character vector
}
\keyword{internal}
