% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{return_eta_indices}
\alias{return_eta_indices}
\title{Convert eta matrix back to original item labels}
\usage{
return_eta_indices(eta, encoding)
}
\arguments{
\item{eta}{Numeric matrix (I x L).}

\item{encoding}{Encoding environment.}
}
\value{
A tibble with original item IDs as row names.
}
\description{
Convert eta matrix back to original item labels
}
\keyword{internal}
