% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{return_pr_indices}
\alias{return_pr_indices}
\title{Convert pr array back to a named list of tibbles}
\usage{
return_pr_indices(pr, encoding)
}
\arguments{
\item{pr}{Numeric 3D array (K x L x R).}

\item{encoding}{Encoding environment.}
}
\value{
A named list of tibbles, one per rating level.
}
\description{
Convert pr array back to a named list of tibbles
}
\keyword{internal}
