% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{return_theta_indices}
\alias{return_theta_indices}
\title{Convert theta matrix back to original user labels}
\usage{
return_theta_indices(theta, encoding)
}
\arguments{
\item{theta}{Numeric matrix (U x K).}

\item{encoding}{Encoding environment.}
}
\value{
A tibble with original user IDs as row names.
}
\description{
Convert theta matrix back to original user labels
}
\keyword{internal}
