% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{structure_folds}
\alias{structure_folds}
\title{Compute the number of items per fold for cross-validation}
\usage{
structure_folds(data, folds)
}
\arguments{
\item{data}{A data.frame with at least two columns (users, items).}

\item{folds}{Integer number of folds.}
}
\value{
Integer: items per fold.
}
\description{
Compute the number of items per fold for cross-validation
}
\keyword{internal}
